Fix bugs such that model can be run from flask UI

+++ assetallocation_UI/aa_web_app/data_import/import_data_from_form.py
@@ -1,5 +1,3 @@
-from decimal import Decimal
-
 from assetallocation_arp.data_etl.dal.data_models.strategy import Times
 
 
@@ -18,8 +16,8 @@ class TimesInputs:
         """
         t = Times(day_of_week=int(self.form.week_day.data), frequency=self.form.frequency.data,
                   leverage_type=self.form.leverage_type.data,
-                  long_signals=[Decimal(i.data) for i in (self.form.sig1_long, self.form.sig2_long, self.form.sig3_long)],
-                  short_signals=[Decimal(i.data) for i in (self.form.sig1_short, self.form.sig2_short, self.form.sig3_short)],
+                  long_signals=[float(i.data) for i in (self.form.sig1_long, self.form.sig2_long, self.form.sig3_long)],
+                  short_signals=[float(i.data) for i in (self.form.sig1_short, self.form.sig2_short, self.form.sig3_short)],
                   time_lag_in_months=int(self.form.time_lag.data),
                   volatility_window=int(self.form.volatility_window.data))
