+++ assetallocation_arp/data_etl/dal/data_models/asset_analytic.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from typing import Union
 from datetime import datetime
 
@@ -8,7 +7,7 @@ from assetallocation_arp.common_libraries.dal_enums.asset_analytic import Catego
 # noinspection PyAttributeOutsideInit
 class AssetAnalytic:
     def __init__(self, asset_ticker: str, category: Union[str, Category],
-                 business_datetime: datetime, value: Decimal) -> None:
+                 business_datetime: datetime, value: float) -> None:
         """AssetAnalytic class to hold data from database"""
         self.asset_ticker = asset_ticker
         self.category = category
@@ -40,9 +39,9 @@ class AssetAnalytic:
         self._category = x if isinstance(x, Category) else Category[x]
 
     @property
-    def value(self) -> Decimal:
+    def value(self) -> float:
         return self._value
 
     @value.setter
-    def value(self, x:  Decimal) -> None:
+    def value(self, x:  float) -> None:
         self._value = x
