+++ tests/assetallocation_arp/data_etl/dal/data_models/test_fund_strategy.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 import mock
 
 from pytest import raises
@@ -8,13 +7,13 @@ from assetallocation_arp.data_etl.dal.data_models.fund_strategy import FundStrat
 
 def test_strategy_name_setter_raises_key_error_invalid_strategy_name():
     with raises(KeyError):
-        FundStrategy('a', 'invalid_strategy_name', 1, Decimal(1))
+        FundStrategy('a', 'invalid_strategy_name', 1, float(1))
 
 
 def test_strategy_name_setter_sets_strategy_name_valid_strategy_name():
     strategy_name = 'times'
 
-    fs = FundStrategy('a', strategy_name, 1, Decimal(1))
+    fs = FundStrategy('a', strategy_name, 1, float(1))
 
     assert fs.strategy_name.name == strategy_name
 
@@ -22,8 +21,8 @@ def test_strategy_name_setter_sets_strategy_name_valid_strategy_name():
 def test_add_fund_strategy_asset_analytic_adds_analytic():
     fsaa_path = 'assetallocation_arp.data_etl.dal.data_models.fund_strategy.FundStrategyAssetAnalytic'
     with mock.patch(fsaa_path, autospec=True) as MockFundStrategyAssetAnalytic:
-        fs = FundStrategy('a', 'times', 1, Decimal(1))
-        fsaa = MockFundStrategyAssetAnalytic('a', 'b', 'c', Decimal(1))
+        fs = FundStrategy('a', 'times', 1, float(1))
+        fsaa = MockFundStrategyAssetAnalytic('a', 'b', 'c', float(1))
 
         fs.add_fund_strategy_asset_analytic(fsaa)
 
@@ -33,8 +32,8 @@ def test_add_fund_strategy_asset_analytic_adds_analytic():
 def test_add_fund_strategy_asset_weight_adds_weight():
     fsaw_path = 'assetallocation_arp.data_etl.dal.data_models.fund_strategy.FundStrategyAssetWeight'
     with mock.patch(fsaw_path, autospec=True) as MockFundStrategyAssetWeight:
-        fs = FundStrategy('a', 'times', 1, Decimal(1))
-        fsaw = MockFundStrategyAssetWeight('a', Decimal(1))
+        fs = FundStrategy('a', 'times', 1, float(1))
+        fsaw = MockFundStrategyAssetWeight('a', float(1))
 
         fs.add_fund_strategy_asset_weight(fsaw)
 
@@ -43,32 +42,32 @@ def test_add_fund_strategy_asset_weight_adds_weight():
 
 def test_category_setter_raises_key_error_invalid_category():
     with raises(KeyError):
-        FundStrategyAssetAnalytic('a', 'invalid_category', 'carry', Decimal(1))
+        FundStrategyAssetAnalytic('a', 'invalid_category', 'carry', float(1))
 
 
 def test_category_setter_sets_category_valid_category():
     category = 'performance'
 
-    fass = FundStrategyAssetAnalytic('a', category, 'carry', Decimal(1))
+    fass = FundStrategyAssetAnalytic('a', category, 'carry', float(1))
 
     assert fass.category.name == category
 
 
 def test_subcategory_setter_raises_key_error_invalid_subcategory_for_performance():
     with raises(KeyError):
-        FundStrategyAssetAnalytic('a', 'performance', 'momentum', Decimal(1))
+        FundStrategyAssetAnalytic('a', 'performance', 'momentum', float(1))
 
 
 def test_subcategory_setter_raises_key_error_invalid_subcategory_for_signal():
     with raises(KeyError):
-        FundStrategyAssetAnalytic('a', 'signal', 'spot', Decimal(1))
+        FundStrategyAssetAnalytic('a', 'signal', 'spot', float(1))
 
 
 def test_subcategory_setter_sets_subcategory_valid_subcategory_for_performance():
     category = 'performance'
     subcategory = 'spot'
 
-    fass = FundStrategyAssetAnalytic('a', category, subcategory, Decimal(1))
+    fass = FundStrategyAssetAnalytic('a', category, subcategory, float(1))
 
     assert fass.subcategory.name == subcategory
 
@@ -77,6 +76,6 @@ def test_subcategory_setter_sets_subcategory_valid_subcategory_for_signal():
     category = 'signal'
     subcategory = 'momentum'
 
-    fass = FundStrategyAssetAnalytic('a', category, subcategory, Decimal(1))
+    fass = FundStrategyAssetAnalytic('a', category, subcategory, float(1))
 
     assert fass.subcategory.name == subcategory
